Game now complete aside from sound effects, music, and mobile support.
Experiencing a NullReferenceException from Edge.cs which I'm not sure how to fix but does not appear to be detrimental to the core functionality of the game or its ability to run correctly, so I'm going to ignore it for now.

I changed the way the death text is displayed, as I found the long horizontal string going off screen to be a bit of an eyesore. It now wraps and stays centered.